rename: base config 파일 네이밍 base 없는거 추가
