+++ apps/admin/eslint.config.mjs
@@ -1,2 +1,2 @@
-import eslintConfig from "../../packages/config/eslint.config.mjs";
+import eslintConfig from "../../packages/config/eslint.config.base.mjs";
 export default eslintConfig;
