+++ apps/web/postcss.config.js
@@ -1,0 +1,3 @@
+const config = require("../../packages/config/postcss.config.base.js");
+
+module.exports = config;
