+++ apps/web/postcss.config.mjs
@@ -1,0 +1,3 @@
+const config = require("../../packages/config/postcss.config.base.mjs");
+
+module.exports = config;
