+++ packages/config/eslint.config.base.mjs
