+++ packages/config/postcss.config.base.js
@@ -1,7 +1,6 @@
-export default {
+module.exports = {
   plugins: {
     tailwindcss: {},
     autoprefixer: {},
   },
-  plugins: ["@tailwindcss/postcss"],
 };
