--- packages/config/tailwind.config.ts
@@ -1,11 +1,0 @@
-import type { Config } from "tailwindcss";
-
-const config: Config = {
-  content: ["../../apps/web/src/**/*.{js,ts,jsx,tsx}", "../../apps/admin/src/**/*.{js,ts,jsx,tsx}"],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
-};
-
-export default config;
